+++ Sources/KeyboardKit/App/KeyboardAppView+RegistrationView.swift
@@ -1,0 +1,21 @@
+//
+//  KeyboardAppView+RegistrationView.swift
+//  KeyboardKit
+//
+//  Created by Daniel Saidi on 2024-08-19.
+//  Copyright © 2024 Daniel Saidi. All rights reserved.
+//
+
+import SwiftUI
+
+extension KeyboardAppView {
+
+    struct RegistrationView: View {
+
+        let app: KeyboardApp
+
+        var body: some View {
+            EmptyView()
+        }
+    }
+}